# cranelift-frontend: Reuse visited block sets in `SSABuilder::can_optimize_var_lookup`