@jameysharp 2022-07-26
Very cool!

If I'm not mis-reading things, `Block` does not need to be dropped (it's just a newtype for u32 and it doesn't implement `Drop`), so `HashSet::clear` is constant-time. In that case, how about just clearing the set before using it, where you're currently `debug_assert`ing that it's empty, and removing all the magic drop-guard machinery?