@fitzgen 2022-07-26
> Also, `can_optimize_var_lookup` is only called from `use_var_nonlocal`, which has a `&mut self`. So you could change `can_optimize_var_lookup` to also take `&mut self` and then get rid of the `RefCell`.

Nice, I didn't notice that!