@fitzgen 2022-07-26
> If I'm not mis-reading things, `Block` does not need to be dropped (it's just a newtype for u32 and it doesn't implement `Drop`), so `HashSet::clear` is constant-time. In that case, how about just clearing the set before using it, where you're currently `debug_assert`ing that it's empty, and removing all the magic drop-guard machinery?

The idea was to empty the set as soon as it isn't used anymore, rather than at the last minute before use, but I guess there isn't really a good reason for that since it doesn't cut down on peak memory usage (which is kind of the whole point of this PR!)