@jameysharp 2022-07-26
Yeah, rather than saying "clear is constant-time", I should have said that it doesn't do anything except mark the table as empty. So it doesn't matter when you do it, as long as nobody else looks at it.